Added color clustering test
Added some code to do color clustering testing, just to see if it works.
It didn't, but I didn't implement it correctly. I just threw all
executions for all the objects in a bit pool and had it do clustering on
that. The correct way to do it would be to have an object-object
distance measure or the clusterer to work with, rather than exec-exec.
Going to work on that next.