Got color clustering test to work with XMeans
ColorClustering test now can correctly cluster the objects using XMeans,
but it only works if XMeans is given the number of clusters to find,
rather than determining that on its own. I need to look up exactly how
XMeans works so I can tweak the parameters so that it will work.