Added ScoredChangeExpWeightedClustering
Added an experimental setup that performs weighted combination of the
contexts. It uses the clustering method of solving the tasks from before
but also does 5-fold cross-validation across the 50 MRT's to train a
weight for each context (just attempts to solve the train tasks with
only that context, and then normalizes all the weights) and then
multiplies the output of the compare method for each context by the
weight, thereby dimenishing the influence of useless contexts.