Changed XMeans to Spectral Clustering
Changed the use of XMeans in the clusterDiffComparator to use Spectral
Clustering instead (it is better able to determine the number of
clusters). Also added two more experiments to test to see if selecting
the top k comparators based on individual performance on a train set or
using boosting on a train set worked better than just simple weighting.
They do not.