Added the correct data
We finished data collection again, but this time the data was
(hopefully) not corrupted. I am committing the save files that have the
new computed features. I also tweaked a couple parameters to fit with
the new data. The old features have been moved to a subfolder. It should
be noted that these may not necessarily be the final features used. The
point clouds collected for look are currently being processed to be used
instead of the ROI color features extracted here. Also, we are currently
developing alternative feature extraction methods, so these may change
in the future.