Ran an actual set of experiments
I changed around the experiment package and subpackages. Since a bunch
of the scoredChangeExps's did not work out (and so aren't going to
actually be used), I moved them to a subpackage called depricated, just
to get them out of the way. I also added an experiment interface and an
experiment controller class that takes a set of experiments and runs
them on lots of different sets of contexts, all multi-threaded. I also
added a results folder where results are output and I added the Tuple
and UnorderedTuple classes.