Reverted back to color histogram
I (manually) reverted the color feature extractor back to doing
histograms over HSV space rather than just averaging pixels in RGB
space. Then I tried some different things to get it to work, but it
still doesn't work. I'm going to go back to using the images and older
version of the color feature extractor.

Also added a new testing class, DistanceMatrixPrinter, which prints out
the distance matrix for a context. Also added MultiJobRunner which takes
a queue of jobs, a class that implements a processJob method, and then
processes the jobs across multiple cores. This is pretty much how I used
MultiThreadRunner before, so it just makes it easier to do. It uses
MultiThreadRunner under the hood to do this.