Fixed bug in color features
Found a bug in the way the color features are computed. Apparently the
8x8 grid that the image was divided into did not take into account the
region of interest that was marked in the image, thus most of the grid
spaces contained all background pixels, meaning the actual foreground
part of the image was only in a few of the cells. Fixed it so that it
only looks at the region of interest, rather than the entire image.